+++ src/main/java/roomescape/controller/api/MemberReservationController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.Auth;
+import roomescape.config.Auth;
 import roomescape.controller.request.LoginMemberInformation;
 import roomescape.controller.request.MemberReservationWebRequest;
 import roomescape.controller.response.MemberReservationWebResponse;
